% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fit_exp_decreasing.R
\name{create_fit_exp_decreasing}
\alias{create_fit_exp_decreasing}
\title{create_fit_exp_decreasing}
\usage{
create_fit_exp_decreasing(
  meancpa,
  mfpga_bin,
  margin,
  meas_no,
  alpha_packet_size
)
}
\arguments{
\item{meancpa}{Sum of the rows from \code{mfpga_bin} and divided by alpha packet size.}

\item{mfpga_bin}{Matrix with each row corresponding to one alpha, each column to an ionization}

\item{margin}{Margin for the calculation of fit, default is 0.01 (one percent).}

\item{meas_no}{Measurement number.}

\item{alpha_packet_size}{Number of alpha particles grouped into one data point, default 100.}
}
\value{
List with entries \code{alph_start}, \code{alph_end}, \code{curve_fit_param_A},
\code{curve_fit_param_B},\code{curve_fit_param_C} and \code{CpT_plot}
}
\description{
This function creates an exponential decreasing fit for measurement data and then
selects a stable region within a margin.
}
\details{
The lowest value of the exponential fit is taken as final value, then the stable region is defined
as the region [alph_start, alph_end]. The \code{alph_start} value is given by where the exponential
decreasing fit crosses the margin around the lowest final value. The function is fitted with
drc::EX.3 and is given by:
\eqn{
f(x) = A + (B-A)(e^{-x/C})
}
}
\examples{
\dontrun{
result_fit1 <- create_fit_exp_decreasing(meancpa = roi_values$meas$meancpa,
                                         mfpga_bin = roi_values$meas$mfpga_bin,
                                         margin = input$roi_margin,
                                         meas_no =  input$roi_meas_no,
                                         alpha_packet_size = input$roi_alpha_packet_size)
roi_values$alph_start <- result_fit1$alph_start
roi_values$alph_end <- result_fit1$alph_end

roi_values$curve_fit_param_A <- result_fit1$curve_fit_param_A
roi_values$curve_fit_param_B <- result_fit1$curve_fit_param_B
roi_values$curve_fit_param_C <- result_fit1$curve_fit_param_C
roi_values$curve_shape <- 1

output$roi_plot_CpT <- result_fit1$CpT_plot
}
}
