% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_meas_data_from_xlsx.R
\name{extract_meas_data_from_xlsx}
\alias{extract_meas_data_from_xlsx}
\title{extract_meas_data_from_xlsx}
\usage{
extract_meas_data_from_xlsx(
  excel_file,
  excel_data_frame,
  meas_no,
  alpha_packet_size = 100
)
}
\arguments{
\item{excel_file}{Excel File Name as a string, example: "Data_Analysis.xlsx" If the file is
in a different folder use the datapath, example: "C:/Users/username/Documents/R_Coding/Excel_File.xlsx"}

\item{excel_data_frame}{Data Frame, obtained using \code{\link{extract_dataFrame_from_xlsx}}.}

\item{meas_no}{Measurement number}

\item{alpha_packet_size}{Number of alpha particles to be grouped together into one data point,
default 100.}
}
\value{
Returns a list list with \code{meancpa}, \code{alpha_non_zero}, \code{counts_pro_alpha},
\code{mfpga_bin},\code{alph_start},\code{alph_end} and \code{alpha_packet_size}.
}
\description{
This function extracts the measurement data of a single measurement in a
dataframe. It uses \code{\link{extract_info_from_xlsx}}.
}
\details{
Excel sheet must contain a sheet named "Info", where in cell B1 the
  THGEM type of the measurements is given and in cell B2 the path to the
  folder with measurement data is given. In the general measurement data folder,
  subfolders for each THGEM should be present with the THGEM name as folder name.

  Folder structure:
  \itemize{
  \item{.../Measurement_Data}
    \itemize{
    \item{em1mm}
    \item{em1p5mm}
    \item{em2mm}
    \item{keramik1p5mm}
  }}
}
\examples{
\dontrun{
roi_values$meas <- extract_meas_data_from_xlsx(excel_file = roi_values$roi_xlsx_file_datapath,
                             excel_data_frame = roi_values$data,
                             meas_no = input$roi_meas_no,
                             alpha_packet_size = input$roi_alpha_packet_size)

meancpa <- roi_values$meas$meancpa
alpha_packet_size <- input$roi_alpha_packet_size

}
}
