% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_CpT.R
\name{make_plot_CpT}
\alias{make_plot_CpT}
\title{make_plot_CpT}
\usage{
make_plot_CpT(meancpa, meas_no, alpha_packet_size)
}
\arguments{
\item{meancpa}{Sum of the rows from \code{mfpga_bin} and divided by alpha packet size.}

\item{meas_no}{Measurement number.}

\item{alpha_packet_size}{Number of alpha particles grouped into one data point, default 100.}
}
\value{
renderPlot object
}
\description{
Creates a plot showing the time evolution of the mean counts per trigger.
}
\examples{
\dontrun{
output$roi_plot_CpT <- make_plot_CpT(meancpa = roi_values$meas$meancpa,
                               meas_no = input$roi_meas_no,
                               alpha_packet_size = input$roi_alpha_packet_size)
}
}
