% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_App_Definition_of_ROI.R
\name{run_App_Definition_of_ROI}
\alias{run_App_Definition_of_ROI}
\title{run_App_Definition_of_ROI}
\usage{
run_App_Definition_of_ROI()
}
\description{
This code launches an app which shows a userinterface for the definition of a stable
region of interest.
}
