% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dataFrame_back_to_xlsx.R
\name{save_dataFrame_back_to_xlsx}
\alias{save_dataFrame_back_to_xlsx}
\title{save_dataFrame_back_to_xlsx}
\usage{
save_dataFrame_back_to_xlsx(excel_file, excel_sheet_name, data)
}
\arguments{
\item{excel_file}{Excel File Name as a string, example: "Data_Analysis.xlsx" If the file is
in a different folder use the datapath, example: "C:/Users/username/Documents/R_Coding/Excel_File.xlsx"}

\item{excel_sheet_name}{Name of Excel Sheet}

\item{data}{Dataframe, which should be saved back to Excel file. Obtained using
\code{\link{extract_dataFrame_from_xlsx}}.}
}
\value{
Executed for its side effects.
}
\description{
Saves dataframe back into Excel Sheet.
}
